Implementing the TiltYieldingFilter
The purpose of this filter is to allow templates that adhere to the Tilt
interface to yield to other templates. Basically, a yields-to chain is
created where the templates sequentially yield to their predecessor in
the list.

For example, given the yields-to chain of [a,b,c] the filter will
process this as a yields-to b and then b yields to c. This causes the
following behavior:
  - c is processed by Tilt, call this c'
  - c' is given as the value of the yield in b
  - b is processed by Tilt, call this b'
  - b' is given as the value of the yield in a
  - a is processed by Tilt, call this a'
  - a' is written as the final output