Implementing the RoyFilter
The RoyFilter compiles files written in the Roy language
(http://roy.brianmckenna.org) to JavaScript as part of the asset
pipeline. This filter depends on the roy-lang gem. It has been added a
dev dependency.

An example using the filter in an Assetfile:

  output "public"

  input "assets" do
    match "*.roy" do
      roy
    end
  end

Also, therubyracer gem was added as a dev depedency because the less
files would not compile without it. Also, the coffee script filter's
spec was modified since the error message on the spec testing invalid
input gave a slighly different message than what the spec was expecting.